+++ lib/packages/wikibase/data-model/tests/unit/SiteLinkListTest.php
@@ -242,8 +242,8 @@ class SiteLinkListTest extends \PHPUnit_Framework_TestCase {
 	public function testAddHasFluentInterface() {
 		$list = new SiteLinkList();
 
-		$list->add( 'enwiki', 'cats' )
-			->add( 'dewiki', 'katzen', array( new ItemId( 'Q1' ) ) );
+		$list->addNewSiteLink( 'enwiki', 'cats' )
+			->addNewSiteLink( 'dewiki', 'katzen', array( new ItemId( 'Q1' ) ) );
 
 		$this->assertTrue( $list->equals( new SiteLinkList( array (
 			new SiteLink( 'enwiki', 'cats' ),
@@ -254,8 +254,8 @@ class SiteLinkListTest extends \PHPUnit_Framework_TestCase {
 	public function testAddObjectHasFluentInterface() {
 		$list = new SiteLinkList();
 
-		$list->addObject( new SiteLink( 'enwiki', 'cats' ) )
-			->addObject( new SiteLink( 'dewiki', 'katzen' ) );
+		$list->addSiteLink( new SiteLink( 'enwiki', 'cats' ) )
+			->addSiteLink( new SiteLink( 'dewiki', 'katzen' ) );
 
 		$this->assertTrue( $list->equals( new SiteLinkList( array (
 			new SiteLink( 'enwiki', 'cats' ),